fix(jans-fido2): cast adoption-rate operand to long before addition

+++ jans-fido2/server/src/main/java/io/jans/fido2/service/metric/Fido2MetricsService.java
@@ -406,7 +406,7 @@ public class Fido2MetricsService {
 
         // Adoption rate: new users against the cumulative population of everyone who has ever
         // registered as of endTime (priorAdopters and newUsers are disjoint by construction).
-        long cumulativeAdopters = priorAdopters.size() + newUsers.size();
+        long cumulativeAdopters = (long) priorAdopters.size() + newUsers.size();
         if (cumulativeAdopters > 0) {
             metrics.put(Fido2MetricsConstants.ADOPTION_RATE, (double) newUsers.size() / cumulativeAdopters);
         } else {
